Incorporate geobuf for passing geojson
According to https://github.com/mapbox/mapbox-gl-js/issues/1504 the
transferring of GeoJSON files between main and worker threads can cause
a bottleneck. We should be able to reduce this bottleneck by converting
and passing the GeoJSON file created in the Worker as a geobuf and
decoding it on the main thread.

This commit is broken because of a possible bug with Webworkify not
pulling in correct dependencies. See
https://github.com/substack/webworkify/issues/14